Reorganize code, add year in review
