+++ src/utils/translate/fr.json
@@ -199,7 +199,8 @@
         "OctoberReport": "Rapport d'Octobre - Covid-19 dépistage et d'essais dans le secteur privé du Canada",
         "PreliminaryReport": "Rapport préliminaire - Covid-19 Dépistage et d'essais dans le secteur privé du Canada",
         "PrePrint": "leçons d'un examen rapide et systématique du début des SRAS-CoV-2 enquêtes sérologiques",
-        "SeptemberReport": "Rapport de Septembre - Covid-19 Dépistage et d'essais dans le secteur privé du Canada"
+        "SeptemberReport": "Rapport de Septembre - Covid-19 Dépistage et d'essais dans le secteur privé du Canada",
+        "YearInReview2020": "SeroTracker 2020 Year in Review"
     },
     "InitInfoModalText": {
         "PartOne": "Bienvenue dans l'onglet Analyse de SeroTracker ! Nous menons un examen systématique des sérosondages du SRAS-CoV-2 et nous utilisons les données provenant de cet examen afin de constituer ce tableau de bord.",
